Fix S3 authN for requests with x-amz-date header
If the 'x-amz-date' is passed with the request, the s3curl was adding both
request Date, and x-amz-date in 'sting to sign' HMAC authentication signature
calculation - which was refused by the Amazon S3 (getting '403 Forbidden').
This fixes the handling, so only the x-amz-date is used with authentication
calculation.

Additional: pass the '-v' to curl if 's3curl.pl --debug' was used.

Test-case:
    mydate=$(date +"%a, %d %b %Y %H:%M:%S %Z")
    s3curl.pl --debug --id=aws --put=myFile http://mybucket.s3.amazonaws.com/myFile -- -H "x-amz-date: $mydate"